fix: rename_variants in absence of discriminants
