+++ tests/discriminants.rs
@@ -631,68 +631,6 @@ fn test_enum_rename_all_with_discriminants() {
 }
 
 #[test]
-fn test_enum_rename_all_rename_variants() {
-    /// Comment for Internal
-    #[derive(Tsify)]
-    #[serde(rename_all = "camelCase")]
-    #[tsify(namespace, rename_variants)]
-    enum Internal {
-        /// Comment for Struct
-        Struct { x: String, y: i32 },
-        /// Comment for EmptyStruct
-        EmptyStruct {},
-        /// Comment for Tuple
-        Tuple(i32, String),
-        /// Comment for EmptyTuple
-        EmptyTuple(),
-        /// Comment for Newtype
-        Newtype(Foo),
-        /// Comment for Unit
-        Unit,
-    }
-
-    let expected = indoc! {r#"
-        type __InternalFoo = Foo;
-        /**
-         * Comment for Internal
-         */
-        declare namespace Internal {
-            /**
-             * Comment for Struct
-             */
-            export type Struct = { struct: { x: string; y: number } };
-            /**
-             * Comment for EmptyStruct
-             */
-            export type EmptyStruct = { emptyStruct: {} };
-            /**
-             * Comment for Tuple
-             */
-            export type Tuple = { tuple: [number, string] };
-            /**
-             * Comment for EmptyTuple
-             */
-            export type EmptyTuple = { emptyTuple: [] };
-            /**
-             * Comment for Newtype
-             */
-            export type Newtype = { newtype: __InternalFoo };
-            /**
-             * Comment for Unit
-             */
-            export type Unit = "unit";
-        }
-
-        /**
-         * Comment for Internal
-         */
-        export type Internal = { [InternalType.Struct]: { x: string; y: number } } | { [InternalType.EmptyStruct]: {} } | { [InternalType.Tuple]: [number, string] } | { [InternalType.EmptyTuple]: [] } | { [InternalType.Newtype]: Foo } | InternalType.Unit;"#
-    };
-
-    assert_eq!(Internal::DECL, expected);
-}
-
-#[test]
 fn test_enum_rename_all_rename_variants_with_discriminants() {
     /// Comment for Internal
     #[derive(Tsify)]
